// comonl-comonlwcl/src/app/modules/comonlapi/model/decodificaGenerica.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';


export interface DecodificaGenerica { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    codDecodifica?: string;
    dsDecodifica?: string;
    idDecodifica?: number;
    idFiltroFacoltativo?: number;
    idComTipoTracciato?: string;
    idComTipoComunicazione?: string;
    flgForma?: string;
    flgStatiUe?: string;
    dataValidita?: Date;
    siglaNazione?: string;
    idCcnl?: string;
    ambitoDecodifica?: string;
    flgAncheNonValidi?: boolean;
}