// comonl-comonlwcl/src/app/modules/comonlapi/model/lavoratore.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Cittadinanza } from './cittadinanza';
import { Comune } from './comune';
import { LivelloStudio } from './livelloStudio';
import { MotivoPermesso } from './motivoPermesso';
import { Questura } from './questura';
import { StatiEsteri } from './statiEsteri';
import { StatusStraniero } from './statusStraniero';


export interface Lavoratore { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    codCapDom?: string;
    codCapRes?: string;
    codiceFiscale?: string;
    cognome?: string;
    dsIndirizzoDom?: string;
    dsIndirizzoRes?: string;
    dtNascita?: Date;
    dtScadenzaPermessoSogg?: Date;
    flgRimborsoRimpatrio?: string;
    flgSistAlloggiativa?: string;
    idLavoratoreSilp?: number;
    nome?: string;
    numeroDocumento?: string;
    sesso?: string;
    statoCivile?: string;
    telDom?: string;
    cittadinanza?: Cittadinanza;
    comuneDom?: Comune;
    comuneRes?: Comune;
    comuneNasc?: Comune;
    livelloStudio?: LivelloStudio;
    motivoPermesso?: MotivoPermesso;
    questura?: Questura;
    statiEsteriDom?: StatiEsteri;
    statiEsteriNasc?: StatiEsteri;
    statiEsteriRes?: StatiEsteri;
    statusStraniero?: StatusStraniero;
}