// comonl-comonlwcl/src/app/modules/comonlapi/model/rapporto.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { CategLavAssObbl } from './categLavAssObbl';
import { CategTirocinante } from './categTirocinante';
import { Ccnl } from './ccnl';
import { Cessazionerl } from './cessazionerl';
import { Comunicazione } from './comunicazione';
import { Cpi } from './cpi';
import { Datore } from './datore';
import { EntePrevidenziale } from './entePrevidenziale';
import { Istat2001livello5 } from './istat2001livello5';
import { LivelloRetribuzione } from './livelloRetribuzione';
import { SedeLavoro } from './sedeLavoro';
import { TipoAttoL68 } from './tipoAttoL68';
import { TipoContratti } from './tipoContratti';
import { TipoEntePromTirocinio } from './tipoEntePromTirocinio';
import { TipoOrario } from './tipoOrario';
import { TipologiaTirocinio } from './tipologiaTirocinio';
import { Trasformazionerl } from './trasformazionerl';
import { Tutore } from './tutore';


export interface Rapporto { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    cfSoggPromotoreTirocinio?: string;
    codiceEntePrev?: string;
    crt?: string;
    dsAttivita?: string;
    dsAttivitaAgricoltura?: string;
    dsSoggPromTirocinio?: string;
    dsVoceTariffa1?: string;
    dsVoceTariffa2?: string;
    dsVoceTariffa3?: string;
    dtApprovazioneCrt?: Date;
    dtCessazione?: Date;
    dtCrt?: Date;
    dtEvento?: Date;
    dtFineMissione?: Date;
    dtFinePeriodoFormativo?: Date;
    dtFineProroga?: Date;
    dtFineRapporto?: Date;
    dtInizioMissione?: Date;
    dtInizioRapporto?: Date;
    dtLegge68?: Date;
    dtTrasformazione?: Date;
    flgAssLegge6899?: string;
    flgAssunzioneObbligatoria?: string;
    flgCcnlApplicato?: string;
    flgDistaccoParziale?: string;
    flgDistaccoSuAziEstera?: string;
    flgLavInMobilita?: string;
    flgLavStagionale?: string;
    flgLavoroAgricoltura?: string;
    flgProsecuzioneDiFatto?: string;
    flgRischioSilicAsbe?: string;
    flgSocioLavoratore?: string;
    flgSommTempoDet?: string;
    flgTirocinio?: string;
    giornateLavPreviste?: number;
    idStoricoRapportoSilp?: number;
    livelloInquadramento?: string;
    numAgenzSomministr?: number;
    numIndennitaSomm?: number;
    numMatricolaLav?: string;
    numOreSettMed?: number;
    numeroAttoLegge68?: string;
    patInail?: string;
    retribuzioneCompenso?: number;
    tipoLavorazione?: string;
    comunicazione?: Comunicazione;
    datoreDistaccatario?: Datore;
    aziUtilizzatrice?: Datore;
    sedeLavoroPrecedente?: SedeLavoro;
    tutore?: Tutore;
    categLavAssObbl?: CategLavAssObbl;
    categTirocinante?: CategTirocinante;
    ccnl?: Ccnl;
    cessazionerl?: Cessazionerl;
    cpi?: Cpi;
    entePrevidenziale?: EntePrevidenziale;
    istatLivello5?: Istat2001livello5;
    livelloRetribuzione?: LivelloRetribuzione;
    tipoAttoL68?: TipoAttoL68;
    tipoContratti?: TipoContratti;
    tipoEntePromTirocinio?: TipoEntePromTirocinio;
    tipologiaTirocinio?: TipologiaTirocinio;
    tipoOrario?: TipoOrario;
    trasformazionerl?: Trasformazionerl;
}